# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateSubscriptionRequest:
    properties:
      end_date:
        example: "2025-12-01T00:00:00Z"
        type: string
      price:
        example: 499
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: "2025-07-01T00:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.SubscriptionResponse:
    properties:
      created_at:
        example: "2025-10-12T19:33:28Z"
        type: string
      end_date:
        example: "2025-12-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      price:
        example: 499
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: "2025-07-01T00:00:00Z"
        type: string
      updated_at:
        example: "2025-10-12T19:33:28Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: "2025-12-31T00:00:00Z"
        type: string
      price:
        example: 599
        type: integer
      service_name:
        example: Yandex Plus
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления подписками пользователей
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      parameters:
      - description: Начало периода (2006-01-02T15:04:05Z)
        in: query
        name: period_start
        required: true
        type: string
      - description: Конец периода (2006-01-02T15:04:05Z)
        in: query
        name: period_end
        required: true
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить суммарную стоимость подписок за период
      tags:
      - subscriptions
swagger: "2.0"